feat: cleaning options repository and validations

+++ app/Http/Controllers/CleaningOptionsController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Helpers\LanguageHelper;
 use App\Repositories\CleaningOptionsRepositoryInterface;
-use App\Models\{ CleaningOption, CleaningOptionTranslation };
+use App\Models\CleaningOption;
 
 class CleaningOptionsController extends Controller
 {
@@ -15,16 +15,17 @@ class CleaningOptionsController extends Controller
     {
         $this->repository = $repository;
     }
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index(Request $request)
-    {
+    {   
         $search = trim($request->s);
         $cleaning_options = $this->repository->all($search);
-        
+
         return view('cleaning-options.index')
             ->with('cleaning_options', $cleaning_options)
             ->with('search', $search);
@@ -50,6 +51,7 @@ class CleaningOptionsController extends Controller
     public function store(Request $request)
     {
         $cleaning_option = $this->repository->create($request);
+        $request->session()->flash('success', __('Record created successfully'));
         return redirect(route('cleaning-options.edit', [$cleaning_option->id]));
     }
 
@@ -60,7 +62,7 @@ class CleaningOptionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function show(CleaningOption $cleaning_option)
-    {
+    {   
         $cleaning_option = $this->repository->find($cleaning_option);        
         return view('cleaning-options.show')->with('cleaning_option', $cleaning_option);
     }
@@ -87,6 +89,7 @@ class CleaningOptionsController extends Controller
     public function update(Request $request, $id)
     {
         $this->repository->update($request, $id);
+        $request->session()->flash('success', __('Record updated successfully'));
         return redirect(route('cleaning-options.edit', [$id]));
     }
 
@@ -96,9 +99,15 @@ class CleaningOptionsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $this->repository->delete($id);
-        return redirect(route('cleaning-options'));
+        if ( $this->repository->canDelete($id) ) {
+            $this->repository->delete($id);
+            $request->session()->flash('success', __('Record deleted successfully'));
+            return redirect(route('cleaning-options'));
+        }
+
+        $request->session()->flash('error', __("This record can't be deleted"));
+        return redirect()->back();
     }
 }
